Modified to use service
