+++ src/app/app.component.html
@@ -1,18 +1,18 @@
 		<section class="todoapp">
-      <app-header [title]='title' (addTodo)="addTodo($event)"></app-header>
-			<!-- This section should be hidden by default and shown when there are todos -->
-			<section class="main" [hidden]="todos.length === 0" >
-				<input class="toggle-all" type="checkbox" [checked]="allChecked" (click)="toggleAllCompleted()">
+      <app-header></app-header>
+			<!-- This section should be hidden by default and shown when there are todoService.todos -->
+			<section class="main" [hidden]="todoService.todos.length === 0" >
+				<input class="toggle-all" type="checkbox" [checked]="todoService.allChecked" (click)="todoService.toggleAllCompleted()">
 				<label for="toggle-all">Mark all as complete</label>
 				<ul class="todo-list">
 					<!-- These are here just to show the structure of the list items -->
 					<!-- List items should get the class `editing` when editing and `completed` when marked as completed -->
-          <li *ngFor="let todo of (todos | filter: filterCondition); let idx = index"  
+          <li *ngFor="let todo of (todoService.todos | filter: todoService.filterCondition); let idx = index"  
              [class.editing]="todo.isEditing" [class.completed]="todo.isCompleted">
               <div class="view">
-                <input class="toggle" type="checkbox" [checked]="todo.isCompleted" (click)="toggleCompleted(todo)">
-                <label (dblclick)="setTodoEditing(todo)">{{ todo.todoNm }}</label>
-                <button class="destroy" (click)="deleteTodo(todo)"></button>
+                <input class="toggle" type="checkbox" [checked]="todo.isCompleted" (click)="todoService.toggleCompleted(todo)">
+                <label (dblclick)="todoService.setTodoEditing(todo)">{{ todo.todoNm }}</label>
+                <button class="destroy" (click)="todoService.deleteTodo(todo)"></button>
               </div>
               <input class="edit" [(ngModel)]="todo.todoNm" (blur)="todo.isEditing=false" (keyup.enter)="todo.isEditing=false">
           </li>
@@ -27,19 +27,19 @@
 					
 				</ul>
 			</section>
-			<!-- This footer should hidden by default and shown when there are todos -->
-      <footer class="footer" *ngIf="todos.length > 0"> 
+			<!-- This footer should hidden by default and shown when there are todoService.todos -->
+      <footer class="footer" *ngIf="todoService.todos.length > 0"> 
         <!-- This should be `0 items left` by default -->
-        <span class="todo-count"><strong>{{ uncompletedCounter }}</strong> item left</span>
+        <span class="todo-count"><strong>{{ todoService.uncompletedCounter }}</strong> item left</span>
         <!-- Remove this if you don't implement routing -->
     
         <!-- 此寫法同下<app-filters 
                    [filterCondition]="filterCondition" 
                    (filterConditionChange)="filterCondition=$event">
            </app-filters> -->
-        <app-filters [(filterCondition)]="filterCondition"></app-filters> 
+        <app-filters></app-filters> 
           <!-- Hidden if no completed items are left ↓ -->
-        <button class="clear-completed" (click)="clearCompleted()">Clear completed</button>
+        <button class="clear-completed" (click)="todoService.clearCompleted()">Clear completed</button>
       </footer>     
     </section>
     <app-footer></app-footer>
