+++ src/app/filters/filters.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { Todo } from '../app.component';
+import { TodosServiceService } from '../todos-service.service';
 
 @Component({
   selector: 'app-filters',
@@ -7,20 +8,9 @@ import { Todo } from '../app.component';
   styleUrls: ['./filters.component.css']
 })
 export class FiltersComponent implements OnInit {
-  _filterCondition: 'all' | 'completed' | 'active';
-  // @Input() todos: Todo[];
-  // @Input() uncompletedCounter: number;
-  @Output() filterConditionChange = new EventEmitter();
-  @Input() get filterCondition() {
-    return this._filterCondition;
-  }
-  set filterCondition(filterCondition: 'all' | 'completed' | 'active') {
-    this._filterCondition = filterCondition;
-    this.filterConditionChange.emit(this._filterCondition);
-  }
   // @Output() clearCompleted = new EventEmitter();
 
-  constructor() { }
+  constructor(public todoService: TodosServiceService) { }
 
   ngOnInit() {
   }
